python-worker/requirements.txt
Pillow==10.3.0
reportlab==4.1.0
fastapi==0.110.0
uvicorn==0.29.0
rembg==2.0.56
opencv-python==4.9.0.80
numpy==1.26.4
imutils==0.5.4
